[MSA-25] refactor(auth): move controllers and DTOs into feature-based packages
