+++ services/auth/internal/controller/http/v1/auth/dto.go
@@ -1,4 +1,4 @@
-package dto
+package auth
 
 // RegisterRequest contains the necessary fields for user registration.
 type RegisterRequest struct {
