+++ services/auth/internal/controller/http/v1/health/dto.go
@@ -1,4 +1,4 @@
-package dto
+package health
 
 // HealthResponse represents the response structure for the health check endpoint.
 // It contains the status of the service (e.g., "ok" if the service is healthy).
