+++ services/auth/internal/controller/http/v1/swagger/swagger.go
@@ -1,4 +1,4 @@
-package v1
+package swagger
 
 import (
 	"github.com/gofiber/fiber/v2"
